LRCI-3181 Move logic Testray Attachments to abstract class BuildTestrayCaseResult
