+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/FunctionalBatchBuildTestrayCaseResult.java
@@ -30,7 +30,8 @@ import java.util.List;
 /**
  * @author Michael Hashimoto
  */
-public class FunctionalBatchBuildTestrayCaseResult extends BatchBuildTestrayCaseResult {
+public class FunctionalBatchBuildTestrayCaseResult
+	extends BatchBuildTestrayCaseResult {
 
 	public FunctionalBatchBuildTestrayCaseResult(
 		TestrayBuild testrayBuild, TopLevelBuild topLevelBuild,
